eliminé la clase employed que no se estaba usando y pasé las variables a User. Monifiqué la clase profile
